refactor: override 메소드 분리
