+++ src/main/java/bowling/domain/frame/NormalFrame.java
@@ -1,6 +1,7 @@
 package bowling.domain.frame;
 
 import bowling.domain.pitch.Pitch;
+import java.util.Optional;
 
 public class NormalFrame extends Frame {
 
@@ -26,5 +27,27 @@ public class NormalFrame extends Frame {
 
     public static NormalFrame of(final Pitch first, final Pitch second) {
         return new NormalFrame(first, second);
+    }
+
+    @Override
+    public boolean isStrike() {
+        return first.isStrike();
+    }
+
+    @Override
+    public boolean isSpare() {
+        return first.getNumber() + second.getNumber() == MAX;
+    }
+
+    @Override
+    public void pitch(final Pitch result) {
+        final Optional<Pitch> first = Optional.ofNullable(this.first);
+
+        if (!first.isPresent()) {
+            this.first = result;
+            return;
+        }
+
+        first.ifPresent(e -> addSecond(result));
     }
 }
